feat: enhance About Us page with Journey and Milestones components, and add background image
