+++ src/app/about-us/page.tsx
@@ -1,6 +1,10 @@
-import AboutUs from '@/components/pages/about-us';
-import React from 'react';
+import { Journey, Milestones } from '@/components/pages/about-us';
 
 export default function AboutUsPage() {
-  return <AboutUs />;
+  return (
+    <div className="max-w-[1920px] mx-auto">
+      <Milestones />
+      <Journey />
+    </div>
+  );
 }
