+++ src/components/pages/about-us/components/Milestones/index.css
